+++ apps/jaspr_pad/lib/providers/project_provider.dart
@@ -74,7 +74,7 @@ final loadedProjectProvider = Provider<AsyncValue<ProjectDataBase>>((ref) {
   }
 
   var storedProject = ref.read(storedProjectProvider);
-  var uri = ref.binding.currentUri;
+  var uri = Uri.parse(ref.binding.currentUrl);
   var queryId = uri.queryParameters['gist'] != null
       ? 'gist-${uri.queryParameters['gist']}'
       : uri.queryParameters['sample'] != null