+++ packages/jaspr/lib/src/browser/app_context.dart
@@ -1,0 +1,6 @@
+import '../framework/framework.dart';
+
+extension AppContext on BuildContext {
+  /// The currently visited url.
+  String get url => binding.currentUrl;
+}